Update HttpKernel to use Authenticate middleware under App namespace
The `Authenticate` middleware is intended to be called in a specific order before applying developer-listed middleware. In 5.7, it was changed to `App\Http\Middleware\Authenticate` but the priority still lists it as living under `Illuminate\Auth\Middleware\Authenticate`.

This proposed fix moves that priority array to `App\Http\Kernel` and changes the reference to the userland class.